Add transparent pixel colour for the pixel art
This is not part of the two colours used to draw with the palette but is
technically needed to allow parts with nothing on for any shapes that
are not square to be able to show what's underneath those pixels.  For
example, the crosshair has a hole in the middle that is transparent and
should show the colour of the pixel underneath it, NOT the dark colour.