Add animations when constructing towers
The code is a mess and needs to be reorganised now but at least I've got
sprite loading working as a proof of concept.  Right now it'll play the
same animation regardless of which tower you place but it's really
groundbreaking for me because I haven't done this kind of animation
loading before where the frame data isn't hardcoded in Go code but comes
from the output of another program.

The graphics for the sprites were made with Aseprite.  I paid for the
full version but it wouldn't start on my computer so I built the dev
version from source.  I tried using aseprite-go but it made a lot of
strange assumptions about the layout of the assets directory and the
structure of the JSON files, which is based on the output of some Lua
scripts which I couldn't install because I couldn't find the Scripts
menu in Aseprite.  After some hours of trying to make it work I gave up
and wrote my own.  It's not so bad although there's at least 1 level of
redundant structs due to the JSON structure.  If I can generalise this a
bit more I might make it into a library.

Now when you place a tower it will appear with an animation, hooray!