Render text to the screen with the "tinier" font
I found a font online that's only 4px tall.  There was a nicer looking
one called tiny that's 5px tall but I couldn't load that as a font face
because it had some error.

I wrote a loadFont() helper just like for music and images that'll load
a TTF font from a file in the embedded FS and return a font face for it.
If there would be more than one font it would probably useful to
parametrise the size of the font used in the function, but this number
seems tricky to figure out.