Hack in animated moving monster sprite
This is by far the worst possible way this could be implemented, globals
everywhere and everything is a mess of long names but it was the
quickest way I could add this to the game to check if these animated
monsters will work.  They do!  Next step will be cleaning up this code
so I can extend it with more monsters!