Add a HUD that displays the amount of money left
With a grid of 7x7 squares there's a 6px space left at the top of the
48px tall screen which I will put in a dark colour and use as a HUD to
display information to the player.  The first thing to get in there is
the money concept previously introduced but not displayed.  The tinier
font doesn't have enough pixels to be able to show a dollar sign $ so
I've put the letter D in its place for now.