Refactor tower loading and include strong-tower sprite
Towers used to have their appearance written in with 0s and 1s by hand,
when the first one got replaced with an image that other stuff was left
there.  It's been cleaned up to some extent now (still needs some work)
but it's enabled me to include the second "strong" tower type again.