Read map waypoint data from a JSON file
I've described the tiles the creeps need to move to as waypoints in a
JSON file.  There associated structs with them (naming got a bit blurred
between ways<->mapdata, too many levels for something so simple maybe)
but it loads now and I've got the spawn point the creeps need to start
moving from.