Draw lasers shooting creeps within 1 tile range of a tower
Such a good feeling that this could be implemented by only editing the
towers.go file :) I added a new Target field to the tower to track which
creep it's attacking.  If a creep's centre point is within the range of
1 7x7 tile in any direction of the tower (including diagonally) then it
sets it as the target.  A line will be drawn from the tower to the
target to represent a laser beam.  Because the creeps currently have no
concept of health the laser has nothing to damage so the creep just
keeps walking but it has a really cool effect as the towers' beams
follow the creeps along the map.