Implement damage dealing from towers
There is now a concept of health (how many hit points a creep has for
taking damage) and also how much damage a tower does when it attacks.
Every tick of the game a tower with a target deals N damage to the
targeted creep until it either dies or moves out of range.

I also had to change the way the towers track their targets to a pointer
reference directly to make it easier to cull dead creeps from the Creeps
array.  If I stayed with them as indexes to the Creeps array then
dropping a creep from the array would cause havock with any towers that
were still tracking it.  Having pointers make it much simpler, and any
reference to that creep can check whether its health is <= and drop the
reference.