Only draw the map image once the game has been loaded
When testing on desktop the files all loaded so fast that this issue
wasn't noticeable but on web this takes longer and the first Draw() call
happens before NewGame() has finished, so there are is not yet any map
to draw.  Moving the map drawing to after the "Loading..." text solves
the problem because that text's logic contains an early return.