Implement very advanced animation for creeps
This uses direction information from the path finding function to
determine which directional animation to use.  If it's moving sideways
it'll used the horizontal frames, otherwise it'll use the vertical
frames.  If the creep is travelling left there's additional logic
involved in the Draw method to flip the image about its centre.