Move spawn points and base off the map
Now instead of creeps appearing out of thin air they start walking on
from off the edge of the map.  Also instead of the game ending as they
enter the last tile near the end it waits until most of the creep's body
is out of sight before the base gets hit and you lose.