Adjust the starting money values to make the game hard
It seems 3 towers is enough to keep small creeps at bay at this speed.
Halving the starting money means you can only place 2 towers to start
with.  Those towers can kill up to 2 creeps before one gets through and
you lose, so I've made the creep loot 50, so that killing two of them
will give you just enough to make a third tower and if you do that fast
enough and put it in the right place you should be able to regain
control of the map.