Don't allow placing buildings on the path
I tried thinking of a more convenient way to do this and gave up.  I
wrote the tile coordinates for every path-covered tile in JSON and then
do some lame maths with it to translate it to rectangles and check
whether the Tower's rectangle would overlap with it or not.  Could
probably have been done with points instead of rectangles and been
simpler, but I am tired at this point...